fix: checkstyle에 맞추어 컨벤션을 수정한다
