+++ api-module/src/test/java/com/romanticpipe/reviewcanvas/domain/shopadmin/application/usecase/ShopAdminUseCaseImplTest.java
@@ -3,8 +3,6 @@ package com.romanticpipe.reviewcanvas.domain.shopadmin.application.usecase;
 import com.romanticpipe.reviewcanvas.admin.service.AdminAuthCreater;
 import com.romanticpipe.reviewcanvas.admin.service.ShopAdminCreator;
 import com.romanticpipe.reviewcanvas.admin.service.ShopAdminValidator;
-import com.romanticpipe.reviewcanvas.service.ReviewDesignReader;
-import com.romanticpipe.reviewcanvas.service.ReviewVisibilityReader;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
@@ -22,10 +20,6 @@ class ShopAdminUseCaseImplTest {
 	ShopAdminCreator shopAdminCreator;
 	@Mock
 	ShopAdminValidator shopAdminValidator;
-	@Mock
-	ReviewVisibilityReader reviewVisibilityReader;
-	@Mock
-	ReviewDesignReader reviewDesignReader;
 	@InjectMocks
 	ShopAdminUseCaseImpl shopAdminUseCase;
