--- domain-module/shopadmin/src/main/java/com/romanticpipe/reviewcanvas/review_property/domain/Shadow.java
@@ -1,5 +1,0 @@
-package com.romanticpipe.reviewcanvas.review_property.domain;
-
-public enum Shadow {
-	NONE, SMALL, MEDIUM, LARGE
-}
