+++ domain-module/shopadmin/src/main/java/com/romanticpipe/reviewcanvas/reviewproperty/domain/FocusAreaLayout.java
@@ -1,4 +1,4 @@
-package com.romanticpipe.reviewcanvas.review_property.domain;
+package com.romanticpipe.reviewcanvas.reviewproperty.domain;
 
 public enum FocusAreaLayout {
 	BEST_REVIEW_TOP, BEST_REVIEW_BOTTOM, BEST_REVIEW_LEFT, BEST_REVIEW_RIGHT
