+++ domain-module/shopadmin/src/main/java/com/romanticpipe/reviewcanvas/reviewproperty/domain/ReviewColumn.java
@@ -1,4 +1,4 @@
-package com.romanticpipe.reviewcanvas.review_property.domain;
+package com.romanticpipe.reviewcanvas.reviewproperty.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embedded;
@@ -16,16 +16,18 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ReviewContainer {
+public class ReviewColumn {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "review_container_id")
+	@Column(name = "review_column_id")
 	private Integer id;
 
 	private String width;
 	@Embedded
 	private Padding padding;
+	@Embedded
+	private Margin margin;
 	private String background;
 	@Embedded
 	private Boarder boarder;
@@ -35,24 +37,27 @@ public class ReviewContainer {
 	private Integer shopAdminId;
 
 	@Builder(access = AccessLevel.PRIVATE)
-	private ReviewContainer(String width, Padding padding, String background, Boarder boarder, Shadow shadow,
-							Integer shopAdminId) {
+	private ReviewColumn(String width, Padding padding, Margin margin, String background, Boarder boarder,
+						 Shadow shadow, Integer shopAdminId) {
 		this.width = width;
 		this.padding = padding;
+		this.margin = margin;
 		this.background = background;
 		this.boarder = boarder;
 		this.shadow = shadow;
 		this.shopAdminId = shopAdminId;
 	}
 
-	public static ReviewContainer create(Integer shopAdminId) {
-		return ReviewContainer.builder()
+	public static ReviewColumn create(Integer shopAdminId) {
+		return ReviewColumn.builder()
 			.width("사이트 Width")
-			.padding(Padding.createDefaultReviewContainer())
+			.padding(Padding.createDefaultReviewColumn())
+			.margin(Margin.createDefaultReviewColumn())
 			.background("blue")
-			.boarder(Boarder.createDefaultReviewContainer())
+			.boarder(Boarder.createDefaultReviewColumn())
 			.shadow(Shadow.NONE)
 			.shopAdminId(shopAdminId)
 			.build();
 	}
+
 }
