+++ domain-module/shopadmin/src/main/java/com/romanticpipe/reviewcanvas/reviewproperty/domain/ReviewContainer.java
@@ -1,4 +1,4 @@
-package com.romanticpipe.reviewcanvas.review_property.domain;
+package com.romanticpipe.reviewcanvas.reviewproperty.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embedded;
@@ -16,18 +16,16 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ReviewColumn {
+public class ReviewContainer {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "review_column_id")
+	@Column(name = "review_container_id")
 	private Integer id;
 
 	private String width;
 	@Embedded
 	private Padding padding;
-	@Embedded
-	private Margin margin;
 	private String background;
 	@Embedded
 	private Boarder boarder;
@@ -37,27 +35,24 @@ public class ReviewColumn {
 	private Integer shopAdminId;
 
 	@Builder(access = AccessLevel.PRIVATE)
-	private ReviewColumn(String width, Padding padding, Margin margin, String background, Boarder boarder,
-						 Shadow shadow, Integer shopAdminId) {
+	private ReviewContainer(String width, Padding padding, String background, Boarder boarder, Shadow shadow,
+							Integer shopAdminId) {
 		this.width = width;
 		this.padding = padding;
-		this.margin = margin;
 		this.background = background;
 		this.boarder = boarder;
 		this.shadow = shadow;
 		this.shopAdminId = shopAdminId;
 	}
 
-	public static ReviewColumn create(Integer shopAdminId) {
-		return ReviewColumn.builder()
+	public static ReviewContainer create(Integer shopAdminId) {
+		return ReviewContainer.builder()
 			.width("사이트 Width")
-			.padding(Padding.createDefaultReviewColumn())
-			.margin(Margin.createDefaultReviewColumn())
+			.padding(Padding.createDefaultReviewContainer())
 			.background("blue")
-			.boarder(Boarder.createDefaultReviewColumn())
+			.boarder(Boarder.createDefaultReviewContainer())
 			.shadow(Shadow.NONE)
 			.shopAdminId(shopAdminId)
 			.build();
 	}
-
 }
