+++ domain-module/shopadmin/src/main/java/com/romanticpipe/reviewcanvas/reviewproperty/domain/ReviewLayoutDesign.java
@@ -1,0 +1,5 @@
+package com.romanticpipe.reviewcanvas.reviewproperty.domain;
+
+public enum ReviewLayoutDesign {
+	BOARD, CARD, TALK
+}
