+++ domain-module/shopadmin/src/main/java/com/romanticpipe/reviewcanvas/reviewproperty/domain/Shadow.java
@@ -1,0 +1,5 @@
+package com.romanticpipe.reviewcanvas.reviewproperty.domain;
+
+public enum Shadow {
+	NONE, SMALL, MEDIUM, LARGE
+}
